requirements.txt
streamlit>=1.31.0
google-genai>=0.1.2
pdfplumber>=0.11.0
PyMuPDF>=1.23.0
pydantic>=2.0.0
